global: phase1 separation of non-clues specific stuff into new libcosmos
I want to better separate generic Linux API related parts from tracing
parts. For this the new "libcosmos" will carry the API related parts
that "libclues" will use to implement the tracing stuff.

This only changes paths and inclues and SConstruct files to get
everything to build again. A couple of naming inconsinstencies and
namespace changes are still missing which will follow in a second step.