cosmos/clues: fix infinite wait() during TermTracer shutdown
The changes in the area of SubProc that introduced the ChildCollector
approach caused a regression in TracedProc. TracedProc "steals" a wait
result from the SubProc machinery. Previously this worked out okay,
because the SubProc was structured more simple. The ChildCollector logic
will enter an infinite wait in this situation, however.

The preliminary solution is somewhat hacky, an API that allows to report
stoled WaitRes events to the ChildCollector.