global: first attempt for proper state modeling
Both ChildTracee and ForeignTracee should work this way now. We also
support forwarding signals to the tracer application now. We properly
stop when seeing SIGSTOP in a tracee and we haven't injected it
ourselves.

There are likely still a ton of problematic states, but this is a good
start, hopefully.