improve tracee exited() handling
It turns out the libcosmos wrapper for the getExitEventMsg() was wrong.
There is more information available from the kernel than I thought at
first. This way we can reliably find out why a tracee is about to exit.

The only piece of heuristic is the detection of the "LOST_TO_EXECVE"
condition, which I attempt by looking at the last trace event that has
been observed.