Tracee: fix strange stopped state after DETACH
The way SIGCONT is currently injected into the tracee causes a strange
behaviour when detaching from the tracee again. The tracee will suddenly
be stuck in stopped state, although no further SIGSTOP was sent to it.
When switching to kill() to send SIGCONT then this strange behavior is
no longer visible. This is also the way `strace` does it.