Introduce Engine class for tracing multiple targets
Currently the tracing loop is found in the Tracee class, which makes it
impossible to trace multiple processes using a single thread.
Multi-threaded tracing could be interesting from a performance point of
view, but also increases complexity substantially in an already complex
area.

Thus introduce an Engine class which manages all tracees and globally
waits for ptrace events. Each event is forwarded to the corresponding
Tracee instance.

Pull out the EventConsumer interface into its own header to remove the
dependency on Tracee.hxx. Also the related Tracee instance now needs to
be reported along with each callback in EventConsumer. We could also use
a pointer member field in EventConsumer for this, but this would make a
later introduction of multi-threading support more complicated.

The detach logic for Tracees also is generalized more with this change.
With the introduction of the Engine class there is also support to stop
tracing at any point in time. Therefore we need to be able to
systematically detach from any present tracees, regardless of their
current state.

The proper cleanup of the SubProc in ChildTracee is not that great yet
and might require further refactoring.