Engine/Tracee: be more robust when Tracees disappear
Avoid uncaught exceptions or exceptions being thrown from destructors
that lead to a core dump. This happens currently e.g. when attaching to
the main thread of a multi-threaded process and a non-main-thread calls
`execve()`.

These scenarios will likely cause more headaches in the future, but
let's patch the worst problems for now at least.