multi-threaded exec: be prepared for main thread not being traced
This case happens when a thread other than the main thread is traced,
and it performs an execve(). The main thread is not visible to Engine,
thus the PID personality change is confusing.

Address this by inspecting the ptrace event in Engine and check whether
the original PID is being traced. This helps to recover this case and
continue tracing under the new personality.

This makes it also necessary to keep track for Tracees via the
ProcessID, not via Tracee*.