Engine: use shared_ptr for Tracee instead of unique_ptr
unique_ptr really isn't the right concept here, since Tracee objects are
passed around a lot and potentially stored in different contexts. Thus
turn it into shared_ptr, the overhead should be bearable given that
we'll rarely trace more than ~100 threads in parallel.

It could be considered to turn all the Tracee& reference in
EventConsumer into TraceePtr now, but at the moment we can do without.

Along with this change the approach to m_main_tracee, which is currently
only needed for the PID. Turning it into a ProcessID simplifies things,
actually.