Engine: remove TODO about inefficient VFORK_DONE Tracee cleanup
While it's true that it will be inefficient, should the client decide to
detach right away, it is the cleanest approach to do it like that. We
can only properly detach if we allocate the new Tracee correctly and
insert it into the Engine's data structures.

Also this situation will likely not occur as often that an optimization
in this area makes sense.