Tracee: handle disappearing Tracees
In some fringe cases, like tracing a single thread of a multi-threaded
process, we can end up with unexplained exit notifications of a Tracee,
without seeing PTHREAD_EVENT_EXIT or anything of the likes.

There is not enough context available to discern what exactly caused
this. Introduce a new callback for this situation and handle it
accordingly in TermTracer.