tools: add helper script for parsing Linux system call tables
We will need to support different architectures at the same time e.g.
for 32-bit emulation. To allow trace applications independent of the
architecture (to the extent possible) we need to be able to smartly
generate headers for the different system call numbers present on
different architectures and abstract them.

This script is a start for this effort. It currently only parses and
displays data, but is supposed to also generate C++ headers in a next
step.