feature: support for X32 system calls
Based on new libcosmos features and the generated syscall nr header for
X32 we can finally detect the X32 ABI and trace it.

For this we need to introduce a new ABI enum, because the Arch provided
on ptrace level does not suffice to identify the ABI a system call
conforms to.