RegisterSet: use ABI-specific data structures
The Tracee logic involving `m_initial_regset` no longer worked when
dealing with 32-bit emulation binaries on X86_64, because RegisterSet
always assumed the native register data structure, while in this context
the I386 set of registers is returned, which was detected due to a size
mismatch.

To fix this we need to model all ABI specific register data structures
and use templates to deal with them. We can no longer rely on the elf.h
data structures, since they only deal with the host architecture seen
during compilation. For a start, this commit introduces data structures
for I386, X86_64 and X32.