SystemCall: fix model for context-dependent parameter setup
There is still a bug in the newSystemCall() approach, since extra
parameters are not cleared in time for the first pass of parameter
parsing, which can lead to errors when e.g. a `struct flock` is still
placed as third parameter to `fcntl()`, while the tracee is actually
performing a completely different `fcntl()` call.

To fix this add a virtual function for preparing a new system call,
which will now perform necessary cleanup steps. newSystemCall() is
renamed to check2ndPass() which more clearly expresses what is happening
here.